change folder name to snake case
