+++ embedchain/config/vector_dbs/elasticsearch_db_config.py
